Creates dedicated ImageManager class with methods to read and write to documentdirectory
